Move user profile to page folder
